Change _all_fields config for normalised records
Change _all_fields in the normalised_summary record definition to be
equivalent to _db_fields, rather than _msg_fields. This is necessary as
_all_fields should be a superset of fields used when converting to and
from the Python record object. It doesn't change what gets written out
to a message, but having it set incorrectly meant that unloading
old-style normalised records from a new-style database was causing an
error when the apel message writer came across the ServiceLevelType
field which it didn't expect at all.