Upgrade fields that match CloudSummary to BIGINT
There was an inconsistency in the CloudRecords table and view where the
durations and some of the other fields used plain INT when the
CloudSummaries table uses BIGINT. There have been records truncated or
rejected due to this. This commit makes the two record types match in
the DB schema (plus SuspendDuration for consistency with the other
durations, even though it's not used in CloudSummaries).